API_v5/requirements.txt
pytest
pytest-html
pytest-xdist
pytest-ordering
pytest-rerunfailures
allure-pytest
jsonpath
colorlog
pyyaml
openpyxl
pymysql
zipfile